fix tests

+++ test/test_distributions.py
@@ -24,19 +24,19 @@ from torchrl.modules.distributions.continuous import SafeTanhTransform
 @pytest.mark.parametrize("div_up", [1, 2])
 @pytest.mark.parametrize("div_down", [1, 2])
 def test_delta(device, div_up, div_down):
-    x = torch.randn(1000000, 4, device=device)
+    x = torch.randn(1000000, 4, device=device, dtype=torch.double)
     d = Delta(x)
     assert d.log_prob(d.mode).shape == x.shape[:-1]
     assert (d.log_prob(d.mode) == float("inf")).all()
 
-    x = torch.randn(1000000, 4, device=device)
+    x = torch.randn(1000000, 4, device=device, dtype=torch.double)
     d = TanhDelta(x, -1 / div_down, 1.0 / div_up, atol=1e-4, rtol=1e-4)
     xinv = d.transforms[0].inv(d.mode)
     assert d.base_dist._is_equal(xinv).all()
     assert d.log_prob(d.mode).shape == x.shape[:-1]
     assert (d.log_prob(d.mode) == float("inf")).all()
 
-    x = torch.randn(1000000, 4, device=device)
+    x = torch.randn(1000000, 4, device=device, dtype=torch.double)
     d = TanhDelta(
         x,
         -torch.ones_like(x) / div_down,
